+++ src/lib/HAPServiceNode.ts
@@ -1,5 +1,4 @@
 import { uuid } from '@homebridge/hap-nodejs'
-import { logger } from '@nrchkb/logger'
 import type { NodeAPI } from 'node-red'
 
 import NRCHKBError from './NRCHKBError'
@@ -7,6 +6,7 @@ import type HAPHostNodeType from './types/HAPHostNodeType'
 import type HAPServiceConfigType from './types/HAPServiceConfigType'
 import type HAPServiceNodeType from './types/HAPServiceNodeType'
 import HostType from './types/HostType'
+import { scopedLogger } from './utils/LogUtils'
 import { NodeStatusUtils } from './utils/NodeStatusUtils'
 
 import buildServiceUtils = require('./utils/ServiceUtils')
@@ -21,7 +21,12 @@ module.exports = (RED: NodeAPI) => {
     const nrchkbConfigCompatibilityOverride = function (
         this: HAPServiceNodeType
     ) {
-        const log = logger('NRCHKB', 'HAPServiceNode', this.config.name, this)
+        const log = scopedLogger(
+            'NRCHKB',
+            'HAPServiceNode',
+            this.config.name,
+            this
+        )
 
         if (this.config.isParent === undefined) {
             log.trace(
@@ -49,7 +54,12 @@ module.exports = (RED: NodeAPI) => {
         this.config = config
         this.name = this.config.name
 
-        const log = logger('NRCHKB', 'HAPServiceNode', this.config.name, this)
+        const log = scopedLogger(
+            'NRCHKB',
+            'HAPServiceNode',
+            this.config.name,
+            this
+        )
 
         this.RED = RED
         this.publishTimers = {}
@@ -94,7 +104,12 @@ module.exports = (RED: NodeAPI) => {
     ) {
         this.config = config
 
-        const log = logger('NRCHKB', 'HAPServiceNode', this.config.name, this)
+        const log = scopedLogger(
+            'NRCHKB',
+            'HAPServiceNode',
+            this.config.name,
+            this
+        )
 
         const ServiceUtils = buildServiceUtils(this)
 
@@ -122,7 +137,12 @@ module.exports = (RED: NodeAPI) => {
     }
 
     const configure = function (this: HAPServiceNodeType) {
-        const log = logger('NRCHKB', 'HAPServiceNode', this.config.name, this)
+        const log = scopedLogger(
+            'NRCHKB',
+            'HAPServiceNode',
+            this.config.name,
+            this
+        )
 
         const ServiceUtils = buildServiceUtils(this)
         const AccessoryUtils = buildAccessoryUtils(this)