+++ src/lib/types/HAPServiceConfigType.ts
@@ -1,5 +1,6 @@
 import type { NodeDef } from 'node-red'
 import type { AdaptiveLightingControllerMode } from '../hap/hap-nodejs'
+import type { NRCHKBLogLevel } from '../utils/LogUtils'
 
 import type CameraConfigType from './CameraConfigType'
 
@@ -15,6 +16,7 @@ type HAPServiceConfigType = NodeDef & {
     serviceName: string
     topic: string
     filter: boolean
+    logLevel?: NRCHKBLogLevel
     manufacturer: string
     model: string
     serialNo: string